requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
motor
passlib[bcrypt]
python-jose[cryptography]
